requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.7
Babel==2.8.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
defusedxml==0.7.0rc1
demjson==2.2.4
distlib==0.3.1
Django==2.2.11
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-filter==2.0.0
django-js-asset==1.2.2
django-otp==0.9.4
django-payu==0.5
django-permalinks==0.1.4
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
django-sendsms==0.3.1
django-thumbor==0.5.7
django-tinymce==3.0.2
django-twilio==0.13.0.post1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
DjangoSEO==1.0
drf-instamojo==0.0.1
drfaddons==0.1.3
filelock==3.0.12
future==0.18.2
gunicorn==19.9.0
httpie==2.2.0
identify==1.4.23
idna==2.9
instamojo-wrapper==1.1
Jinja2==2.10.3
joblib==0.15.1
kiwisolver==1.2.0
libthumbor==1.3.2
MarkupSafe==1.1.1
matplotlib==3.3.0
nodeenv==1.4.0
nose==1.3.7
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
passlib==1.7.2
phonenumbers==8.11.2
Pillow==7.1.2
pre-commit==2.6.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
qrcode==6.1
razorpay==1.2.0
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.22
scipy==1.4.1
serializers==0.2.4
six==1.14.0
sklearn==0.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
threadpoolctl==2.0.0
toml==0.10.1
twilio==6.39.0
urllib3==1.25.9
virtualenv==20.0.27
xlrd==1.2.0
xmltodict==0.12.0
